+++ tests/unit/test_processors/conftest.py
@@ -4,14 +4,11 @@
 #
 # Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific language governing permissions and limitations under the License.
 
-def get_metrics_mocks(mocker):
-    mock_set_meter_provider = mocker.Mock()
-    mock_metrics = mocker.patch(
-        "solarwinds_apm.configurator.metrics",
+import pytest
+
+@pytest.fixture(name="mock_meter_manager")
+def mock_meter_manager(mocker):
+    return mocker.patch(
+        "solarwinds_apm.apm_meter_manager.SolarWindsMeterManager",
     )
-    mock_metrics.configure_mock(
-        **{
-            "set_meter_provider": mock_set_meter_provider
-        }
-    )
-    return mock_metrics
+