+++ tests/unit/test_processors/test_otlp_metrics_processor.py
@@ -6,108 +6,139 @@
 
 from os import environ
 
+from solarwinds_apm.apm_meter_manager import SolarWindsMeterManager
+from solarwinds_apm.apm_noop import SolarWindsMeterManager as NoopMeterManager
+
 from solarwinds_apm.trace import SolarWindsOTLPMetricsSpanProcessor
 from solarwinds_apm.trace.tnames import TransactionNames
 
 class TestSolarWindsOTLPMetricsSpanProcessor:
 
-    def test__init(self, mocker):
-        mock_tx_mgr = mocker.Mock()
-        mock_meters = mocker.Mock()
+    def get_mock_apm_config(
+        self,
+        mocker,
+        outer_txn_retval="unused",
+        lambda_function_name="unused",
+        inner_retval=True,
+    ):
         mock_apm_config = mocker.Mock()
+        mock_get_inner = mocker.Mock(return_value=inner_retval)
+        mock_inner = mocker.Mock()
+        mock_inner.configure_mock(
+            **{
+                "get": mock_get_inner,
+            }
+        )
+        def outer_side_effect(cnf_key):
+            if cnf_key == "experimental":
+                return mock_inner
+            # We mock this assuming processor does
+            # one outer `get` of 'transaction_name'
+            else:
+                return outer_txn_retval
+
+        mock_get_outer = mocker.Mock(
+            side_effect=outer_side_effect,
+        )
         mock_apm_config.configure_mock(
             **{
-                "service_name": "foo-service-name",
-                # one call to get "transaction_name"
-                "get": mocker.Mock(return_value="foo-env-txn-name"),
-                "lambda_function_name": "foo-lambda-name",
+                "service_name": "foo-service",
+                "get": mock_get_outer,
+                "lambda_function_name": lambda_function_name,
             }
+        )
+        return mock_apm_config
+
+    def test__init(self, mocker, mock_meter_manager):
+        mock_tx_mgr = mocker.Mock()
+        mock_oboe_api = mocker.Mock()
+        mock_apm_config = self.get_mock_apm_config(
+            mocker,
+            "foo-env-txn-name",
+            "foo-lambda-name",
         )
         processor = SolarWindsOTLPMetricsSpanProcessor(
             mock_tx_mgr,
             mock_apm_config,
-            mock_meters,
+            mock_oboe_api,
         )
         assert processor.apm_txname_manager == mock_tx_mgr
-        assert processor.apm_meters == mock_meters
-        assert processor.service_name == "foo-service-name"
+        assert processor.service_name == "foo-service"
         assert processor.env_transaction_name == "foo-env-txn-name"
         assert processor.lambda_function_name == "foo-lambda-name"
+        assert isinstance(processor.apm_meters, SolarWindsMeterManager)
 
-    def test_calculate_otlp_transaction_name_custom(self, mocker):
-        mock_apm_config = mocker.Mock()
-        mock_apm_config.configure_mock(
-            **{
-                # one call to get "transaction_name"
-                "get": mocker.Mock(return_value="unused"),
-                "lambda_function_name": "unused",
-            }
+    def test__init_not_experimental(self, mocker, mock_meter_manager):
+        mock_tx_mgr = mocker.Mock()
+        mock_oboe_api = mocker.Mock()
+        mock_apm_config = self.get_mock_apm_config(
+            mocker,
+            "foo-env-txn-name",
+            "foo-lambda-name",
+            False,
         )
+        processor = SolarWindsOTLPMetricsSpanProcessor(
+            mock_tx_mgr,
+            mock_apm_config,
+            mock_oboe_api,
+        )
+        assert processor.apm_txname_manager == mock_tx_mgr
+        assert processor.service_name == "foo-service"
+        assert processor.env_transaction_name == "foo-env-txn-name"
+        assert processor.lambda_function_name == "foo-lambda-name"
+        assert isinstance(processor.apm_meters, NoopMeterManager)
 
+    def test_calculate_otlp_transaction_name_custom(self, mocker, mock_meter_manager):
+        mock_apm_config = self.get_mock_apm_config(mocker)
         tnames = TransactionNames(
             "unused",
             "unused",
             "foo-custom-name",
         )
-
         assert "foo-custom-name" == SolarWindsOTLPMetricsSpanProcessor(
             mocker.Mock(),
             mock_apm_config,
             mocker.Mock()
         ).calculate_otlp_transaction_name(tnames)
 
-    def test_calculate_otlp_transaction_name_env_trans(self, mocker):
-        mock_apm_config = mocker.Mock()
-        mock_apm_config.configure_mock(
-            **{
-                # one call to get "transaction_name"
-                "get": mocker.Mock(return_value="foo-env-trans-name"),
-                "lambda_function_name": "unused",
-            }
+    def test_calculate_otlp_transaction_name_env_trans(self, mocker, mock_meter_manager):
+        mock_apm_config = self.get_mock_apm_config(
+            mocker,
+            "foo-env-trans-name",
         )
-
         tnames = TransactionNames(
             "unused",
             "unused",
             None,
         )
-
         assert "foo-env-trans-name" == SolarWindsOTLPMetricsSpanProcessor(
             mocker.Mock(),
             mock_apm_config,
             mocker.Mock()
         ).calculate_otlp_transaction_name(tnames)
 
-    def test_calculate_otlp_transaction_name_env_lambda(self, mocker):
-        mock_apm_config = mocker.Mock()
-        mock_apm_config.configure_mock(
-            **{
-                # one call to get "transaction_name"
-                "get": mocker.Mock(return_value=None),
-                "lambda_function_name": "foo-lambda-name",
-            }
+    def test_calculate_otlp_transaction_name_env_lambda(self, mocker, mock_meter_manager):
+        mock_apm_config = self.get_mock_apm_config(
+            mocker,
+            outer_txn_retval=None,
+            lambda_function_name="foo-lambda-name",
         )
-
         tnames = TransactionNames(
             "unused",
             "unused",
             None,
         )
-
         assert "foo-lambda-name" == SolarWindsOTLPMetricsSpanProcessor(
             mocker.Mock(),
             mock_apm_config,
             mocker.Mock()
         ).calculate_otlp_transaction_name(tnames)
 
-    def test_calculate_otlp_transaction_name_default(self, mocker):
-        mock_apm_config = mocker.Mock()
-        mock_apm_config.configure_mock(
-            **{
-                # one call to get "transaction_name"
-                "get": mocker.Mock(return_value=None),
-                "lambda_function_name": None,
-            }
+    def test_calculate_otlp_transaction_name_default(self, mocker, mock_meter_manager):
+        mock_apm_config = self.get_mock_apm_config(
+            mocker,
+            outer_txn_retval=None,
+            lambda_function_name=None,
         )
 
         tnames = TransactionNames(
@@ -160,12 +191,7 @@ class TestSolarWindsOTLPMetricsSpanProcessor:
         )
         mock_get_http_status_code.configure_mock(return_value="foo-code")
 
-        mock_apm_config = mocker.Mock()
-        mock_apm_config.configure_mock(
-            **{
-                "service_name": "foo-service"
-            }
-        )
+        mock_apm_config = self.get_mock_apm_config(mocker)
 
         mock_txname_manager = mocker.Mock()
         mock_set = mocker.Mock()
@@ -188,9 +214,7 @@ class TestSolarWindsOTLPMetricsSpanProcessor:
             }
         )
 
-        mock_response_time = mocker.PropertyMock()
-        mock_meters = mocker.Mock()
-        type(mock_meters).response_time = mock_response_time
+        mock_oboe_api = mocker.Mock()
 
         mock_basic_span = mocker.Mock()
         mock_basic_span.configure_mock(
@@ -200,17 +224,28 @@ class TestSolarWindsOTLPMetricsSpanProcessor:
                 },
             }
         )
-        
+
+        mock_response_time = mocker.Mock()
+        mock_response_time.record = mocker.Mock()
+        mock_mm = mocker.Mock()
+        type(mock_mm).response_time = mock_response_time
+        mocker.patch(
+            "solarwinds_apm.trace.otlp_metrics_processor.SolarWindsMeterManager",
+            return_value=mock_mm,
+        )
+
         return mock_txname_manager, \
             mock_apm_config, \
-            mock_meters, \
+            mock_oboe_api, \
+            mock_response_time, \
             mock_basic_span
 
     def test_on_end_valid_local_parent_span(self, mocker):
         """Only scenario to skip OTLP metrics generation (not entry span)"""
         mock_txname_manager, \
             mock_apm_config, \
-            mock_meters, \
+            mock_oboe_api, \
+            mock_response_time, \
             _ = self.patch_for_on_end(
                 mocker,
             )
@@ -231,15 +266,17 @@ class TestSolarWindsOTLPMetricsSpanProcessor:
         processor = SolarWindsOTLPMetricsSpanProcessor(
             mock_txname_manager,
             mock_apm_config,
-            mock_meters,
+            mock_oboe_api,
         )
         processor.on_end(mock_span)
-        mock_meters.response_time.record.assert_not_called()
+
+        mock_response_time.record.assert_not_called()
 
     def test_on_end_valid_remote_parent_span(self, mocker):
         mock_txname_manager, \
             mock_apm_config, \
-            mock_meters, \
+            mock_oboe_api, \
+            mock_response_time, \
             _ = self.patch_for_on_end(
                 mocker,
             )
@@ -253,23 +290,40 @@ class TestSolarWindsOTLPMetricsSpanProcessor:
         )
         mock_span.configure_mock(
             **{
-                "parent": mock_parent
+                "parent": mock_parent,
+                "attributes": {
+                    "http.method": "foo-method"
+                }
             }
         )
 
         processor = SolarWindsOTLPMetricsSpanProcessor(
             mock_txname_manager,
             mock_apm_config,
-            mock_meters,
+            mock_oboe_api,
         )
         processor.on_end(mock_span)
 
-        mock_meters.response_time.record.assert_called_once()
+        mock_response_time.record.assert_called_once()
+        mock_response_time.record.assert_has_calls(
+            [
+                mocker.call(
+                    amount=123,
+                    attributes={
+                        'sw.is_error': 'true',
+                        'http.status_code': 'foo-code',
+                        'http.method': 'foo-method',
+                        'sw.transaction': 'foo'
+                    }
+                )
+            ]
+        )
 
-    def test_on_end_invalid_remote_parent_span(self, mocker):
+    def test_on_end_invalid_remote_parent_span(self, mocker, mock_meter_manager):
         mock_txname_manager, \
             mock_apm_config, \
-            mock_meters, \
+            mock_oboe_api, \
+            mock_response_time, \
             _ = self.patch_for_on_end(
                 mocker,
             )
@@ -283,23 +337,40 @@ class TestSolarWindsOTLPMetricsSpanProcessor:
         )
         mock_span.configure_mock(
             **{
-                "parent": mock_parent
+                "parent": mock_parent,
+                "attributes": {
+                    "http.method": "foo-method"
+                }
             }
         )
 
         processor = SolarWindsOTLPMetricsSpanProcessor(
             mock_txname_manager,
             mock_apm_config,
-            mock_meters,
+            mock_oboe_api,
         )
         processor.on_end(mock_span)
 
-        mock_meters.response_time.record.assert_called_once()
+        mock_response_time.record.assert_called_once()
+        mock_response_time.record.assert_has_calls(
+            [
+                mocker.call(
+                    amount=123,
+                    attributes={
+                        'sw.is_error': 'true',
+                        'http.status_code': 'foo-code',
+                        'http.method': 'foo-method',
+                        'sw.transaction': 'foo'
+                    }
+                )
+            ]
+        )
 
-    def test_on_end_invalid_local_parent_span(self, mocker):
+    def test_on_end_invalid_local_parent_span(self, mocker, mock_meter_manager):
         mock_txname_manager, \
             mock_apm_config, \
-            mock_meters, \
+            mock_oboe_api, \
+            mock_response_time, \
             _ = self.patch_for_on_end(
                 mocker,
             )
@@ -313,46 +384,80 @@ class TestSolarWindsOTLPMetricsSpanProcessor:
         )
         mock_span.configure_mock(
             **{
-                "parent": mock_parent
+                "parent": mock_parent,
+                "attributes": {
+                    "http.method": "foo-method"
+                }
             }
         )
 
         processor = SolarWindsOTLPMetricsSpanProcessor(
             mock_txname_manager,
             mock_apm_config,
-            mock_meters,
+            mock_oboe_api,
         )
         processor.on_end(mock_span)
 
-        mock_meters.response_time.record.assert_called_once()
+        mock_response_time.record.assert_called_once()
+        mock_response_time.record.assert_has_calls(
+            [
+                mocker.call(
+                    amount=123,
+                    attributes={
+                        'sw.is_error': 'true',
+                        'http.status_code': 'foo-code',
+                        'http.method': 'foo-method',
+                        'sw.transaction': 'foo'
+                    }
+                )
+            ]
+        )
 
-    def test_on_end_missing_parent(self, mocker):
+    def test_on_end_missing_parent(self, mocker, mock_meter_manager):
         mock_txname_manager, \
             mock_apm_config, \
-            mock_meters, \
+            mock_oboe_api, \
+            mock_response_time, \
             _ = self.patch_for_on_end(
                 mocker,
             )
         mock_span = mocker.Mock()
         mock_span.configure_mock(
             **{
-                "parent": None
+                "parent": None,
+                "attributes": {
+                    "http.method": "foo-method"
+                }
             }
         )
 
         processor = SolarWindsOTLPMetricsSpanProcessor(
             mock_txname_manager,
             mock_apm_config,
-            mock_meters,
+            mock_oboe_api,
         )
         processor.on_end(mock_span)
 
-        mock_meters.response_time.record.assert_called_once()
+        mock_response_time.record.assert_called_once()
+        mock_response_time.record.assert_has_calls(
+            [
+                mocker.call(
+                    amount=123,
+                    attributes={
+                        'sw.is_error': 'true',
+                        'http.status_code': 'foo-code',
+                        'http.method': 'foo-method',
+                        'sw.transaction': 'foo'
+                    }
+                )
+            ]
+        )
 
-    def test_on_end_missing_txn_name(self, mocker):
+    def test_on_end_missing_txn_name(self, mocker, mock_meter_manager):
         mock_txname_manager, \
             mock_apm_config, \
-            mock_meters, \
+            mock_oboe_api, \
+            mock_response_time, \
             mock_basic_span = self.patch_for_on_end(
                 mocker,
                 get_retval=None,
@@ -361,16 +466,17 @@ class TestSolarWindsOTLPMetricsSpanProcessor:
         processor = SolarWindsOTLPMetricsSpanProcessor(
             mock_txname_manager,
             mock_apm_config,
-            mock_meters,
+            mock_oboe_api,
         )
         processor.on_end(mock_basic_span)
 
-        mock_meters.response_time.record.assert_not_called()
+        mock_oboe_api.response_time.record.assert_not_called()
 
-    def test_on_end_txn_name_wrong_type(self, mocker):
+    def test_on_end_txn_name_wrong_type(self, mocker, mock_meter_manager):
         mock_txname_manager, \
             mock_apm_config, \
-            mock_meters, \
+            mock_oboe_api, \
+            mock_response_time, \
             mock_basic_span = self.patch_for_on_end(
                 mocker,
                 get_retval="some-str",
@@ -379,16 +485,17 @@ class TestSolarWindsOTLPMetricsSpanProcessor:
         processor = SolarWindsOTLPMetricsSpanProcessor(
             mock_txname_manager,
             mock_apm_config,
-            mock_meters,
+            mock_oboe_api,
         )
         processor.on_end(mock_basic_span)
 
-        mock_meters.response_time.record.assert_not_called()
+        mock_oboe_api.response_time.record.assert_not_called()
 
-    def test_on_end_is_span_http_has_error(self, mocker):
+    def test_on_end_is_span_http_has_error(self, mocker, mock_meter_manager):
         mock_txname_manager, \
             mock_apm_config, \
-            mock_meters, \
+            mock_oboe_api, \
+            mock_response_time, \
             mock_basic_span = self.patch_for_on_end(
                 mocker,
                 has_error=True,
@@ -398,24 +505,30 @@ class TestSolarWindsOTLPMetricsSpanProcessor:
         processor = SolarWindsOTLPMetricsSpanProcessor(
             mock_txname_manager,
             mock_apm_config,
-            mock_meters,
+            mock_oboe_api,
         )
         processor.on_end(mock_basic_span)
 
-        mock_meters.response_time.record.assert_called_once_with(
-            amount=123,
-            attributes={
-                'sw.is_error': 'true',
-                'http.status_code': 'foo-code',
-                'http.method': 'foo-method',
-                'sw.transaction': 'foo'
-            }
+        mock_response_time.record.assert_called_once()
+        mock_response_time.record.assert_has_calls(
+            [
+                mocker.call(
+                    amount=123,
+                    attributes={
+                        'sw.is_error': 'true',
+                        'http.status_code': 'foo-code',
+                        'http.method': 'foo-method',
+                        'sw.transaction': 'foo'
+                    }
+                )
+            ]
         )
 
-    def test_on_end_is_span_http_not_has_error(self, mocker):
+    def test_on_end_is_span_http_not_has_error(self, mocker, mock_meter_manager):
         mock_txname_manager, \
             mock_apm_config, \
-            mock_meters, \
+            mock_oboe_api, \
+            mock_response_time, \
             mock_basic_span = self.patch_for_on_end(
                 mocker,
                 has_error=False,
@@ -425,24 +538,30 @@ class TestSolarWindsOTLPMetricsSpanProcessor:
         processor = SolarWindsOTLPMetricsSpanProcessor(
             mock_txname_manager,
             mock_apm_config,
-            mock_meters,
+            mock_oboe_api,
         )
         processor.on_end(mock_basic_span)
 
-        mock_meters.response_time.record.assert_called_once_with(
-            amount=123,
-            attributes={
-                'sw.is_error': 'false',
-                'http.status_code': 'foo-code',
-                'http.method': 'foo-method',
-                'sw.transaction': 'foo'
-            }
+        mock_response_time.record.assert_called_once()
+        mock_response_time.record.assert_has_calls(
+            [
+                mocker.call(
+                    amount=123,
+                    attributes={
+                        'sw.is_error': 'false',
+                        'http.status_code': 'foo-code',
+                        'http.method': 'foo-method',
+                        'sw.transaction': 'foo'
+                    }
+                )
+            ]
         )
 
-    def test_on_end_not_is_span_http_has_error(self, mocker):
+    def test_on_end_not_is_span_http_has_error(self, mocker, mock_meter_manager):
         mock_txname_manager, \
             mock_apm_config, \
-            mock_meters, \
+            mock_oboe_api, \
+            mock_response_time, \
             mock_basic_span = self.patch_for_on_end(
                 mocker,
                 has_error=True,
@@ -452,22 +571,28 @@ class TestSolarWindsOTLPMetricsSpanProcessor:
         processor = SolarWindsOTLPMetricsSpanProcessor(
             mock_txname_manager,
             mock_apm_config,
-            mock_meters,
+            mock_oboe_api,
         )
         processor.on_end(mock_basic_span)
 
-        mock_meters.response_time.record.assert_called_once_with(
-            amount=123,
-            attributes={
-                'sw.is_error': 'true',
-                'sw.transaction': 'foo'
-            }
+        mock_response_time.record.assert_called_once()
+        mock_response_time.record.assert_has_calls(
+            [
+                mocker.call(
+                    amount=123,
+                    attributes={
+                        'sw.is_error': 'true',
+                        'sw.transaction': 'foo'
+                    }
+                )
+            ]
         )
 
-    def test_on_end_not_is_span_http_not_has_error(self, mocker):
+    def test_on_end_not_is_span_http_not_has_error(self, mocker, mock_meter_manager):
         mock_txname_manager, \
             mock_apm_config, \
-            mock_meters, \
+            mock_oboe_api, \
+            mock_response_time, \
             mock_basic_span = self.patch_for_on_end(
                 mocker,
                 has_error=False,
@@ -478,14 +603,19 @@ class TestSolarWindsOTLPMetricsSpanProcessor:
         processor = SolarWindsOTLPMetricsSpanProcessor(
             mock_txname_manager,
             mock_apm_config,
-            mock_meters,
+            mock_oboe_api,
         )
         processor.on_end(mock_basic_span)
 
-        mock_meters.response_time.record.assert_called_once_with(
-            amount=123,
-            attributes={
-                'sw.is_error': 'false',
-                'sw.transaction': 'foo'
-            }
+        mock_response_time.record.assert_called_once()
+        mock_response_time.record.assert_has_calls(
+            [
+                mocker.call(
+                    amount=123,
+                    attributes={
+                        'sw.is_error': 'false',
+                        'sw.transaction': 'foo'
+                    }
+                )
+            ]
         )