+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased](https://github.com/solarwinds/apm-python/compare/rel-1.1.1...HEAD)
 
+### Changed
+- Updated OTLP metrics components initialization ([#289](https://github.com/solarwinds/apm-python/pull/289))
+
 
 ## [1.1.1](https://github.com/solarwinds/apm-python/releases/tag/rel-1.1.1) - 2024-01-18